requirements.txt
robotframework
robotframework-pythonlibcore
robotframework-seleniumlibrary
